+++ taxel-xml/src/lib.rs
@@ -9,70 +9,10 @@ pub use crate::csv::{
 };
 use log::warn;
 pub use quick_xml::{Reader, Writer};
-use serde::{Deserialize, Serialize};
-use std::{collections::HashMap, fmt};
+use std::collections::HashMap;
 pub use xbrl::XbrlElement;
 pub use xml::extract_tag_values;
 pub use xml::remove_formatting;
-
-struct Attribute<'a> {
-    key: &'a str,
-    value: &'a str,
-}
-
-const NIL_ATTRIBUTE: Attribute = Attribute {
-    key: "xsi:nil",
-    value: "true",
-};
-
-const DECIMALS_2: Attribute = Attribute {
-    key: "decimals",
-    value: "2",
-};
-
-#[derive(Debug, PartialEq, Clone)]
-/// A struct representing the supported taxonomies.
-pub enum Taxonomy {
-    /// The Global Common Document (GCD) financial reporting taxonomy.
-    Gcd,
-    /// The Generally Accepted Accounting Principles (GAAP) - current/invested
-    /// (CI) - financial reporting taxonomy.
-    GaapCi,
-}
-
-impl Taxonomy {
-    fn as_str(&self) -> &str {
-        match self {
-            Self::Gcd => "gcd",
-            Self::GaapCi => "gaap-ci",
-        }
-    }
-}
-
-impl fmt::Display for Taxonomy {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        let taxonomy = match self {
-            Self::Gcd => "gcd",
-            Self::GaapCi => "gaap-ci",
-        };
-
-        write!(f, "{}", taxonomy)
-    }
-}
-
-#[derive(Debug, Serialize, Deserialize)]
-struct CsvRow {
-    #[serde(rename = "ebilanz_key")]
-    key: String,
-    #[serde(rename = "ebilanz_value")]
-    value: Option<String>,
-}
-
-impl CsvRow {
-    pub fn new(key: String, value: Option<String>) -> Self {
-        Self { key, value }
-    }
-}
 
 #[derive(Debug, PartialEq)]
 pub struct Tag {
@@ -89,14 +29,14 @@ impl Tag {
     }
 }
 
+#[derive(Debug, PartialEq)]
+pub struct TargetTags(HashMap<String, Option<String>>);
+
 impl Default for TargetTags {
     fn default() -> Self {
         Self::new()
     }
 }
-
-#[derive(Debug, PartialEq)]
-pub struct TargetTags(HashMap<String, Option<String>>);
 
 impl TargetTags {
     pub fn new() -> Self {
